Reorganise files and made a QCasheModel that will be used to view large files, performance improvements
